Prepare the partitions for reduce operation
When the data has been properly get by the last operation, the job_queue will wait for all the partitions from the dispatchers. 
Then it splits the data into partitions that are key, ListOfValues tuples for the reduce operation.
If the partitions are more than the number of key, the input of the operation is empty.