Coordinator now handle to resend works + sleepy workers
In case of a crash by a worker, now the coordinator, resend to the queue the pending job. 
It is now implemented a map #{socket => Job} to keep track of pending jobs and resend them quickly to the queue.
Also added two shell commands to quickly recall the erl terminal, clear the shell and in case of c (coordinator) re compile with compile.sh script
Also, the workers now are slowed down by a sleep for testing reasons